Simplify path matching logic.
